tools/make-deb/pyproject.toml
[project]
name = "make-deb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.5",
]
